Rename isAllTimeDateFilter to isAllTimeDateFilterOption in SPI
This is to avoid name clashes with the newly added isAllTimeDateFilter in sdk-model.

Also, misnamed (and therefore not run) test files fixed.

JIRA: RAIL-2633
